# swagger.json
{
	"components": {
...
		"schemas": {
			"IResourceObject": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"type": "string"
				}
			},
			"IQueryResults": {
				"properties": {
					"head": {
						"properties": {
							"vars": {
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						},
						"required": [
							"vars"
						],
						"type": "object"
					},
					"results": {
						"properties": {
							"bindings": {
								"properties": {},
								"additionalProperties": {
									"items": {
										"properties": {
											"value": {
												"type": "string"
											},
											"type": {
												"type": "string"
											}
										},
										"required": [
											"value",
											"type"
										],
										"type": "object"
									},
									"type": "array"
								},
								"type": "object"
							}
						},
						"required": [
							"bindings"
						],
						"type": "object"
					}
				},
				"required": [
					"head",
					"results"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IReturnProject": {
				"properties": {
					"metadata": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"graphs": {
						"$ref": "#/components/schemas/IResourceObject"
					},
					"documents": {
						"$ref": "#/components/schemas/IResourceObject"
					},
					"permissions": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"queryResults": {
						"$ref": "#/components/schemas/IQueryResults"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"metadata",
					"id",
					"graphs",
					"documents"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateProject": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"open": {
						"type": "boolean"
					}
				},
				"required": [
					"title",
					"description",
					"open"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IReturnResource": {
				"properties": {
					"uri": {
						"type": "string"
					},
					"metadata": {
						"type": "string"
					},
					"data": {
						"anyOf": [
							{
								"type": "string",
								"format": "byte"
							},
							{
								"type": "string"
							}
						]
					},
					"results": {
						"$ref": "#/components/schemas/IQueryResults"
					}
				},
				"required": [
					"uri"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "lbd-middleware",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/lbd": {
			"get": {
				"operationId": "GetAllProjects",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IReturnProject"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"reason": {
											"properties": {},
											"additionalProperties": {},
											"type": "object"
										}
									},
									"required": [
										"message",
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Retrieves all the projects associated with the authenticated user. The user is authenticated via a Bearer token sent along with the request as a header \"Authorization: Bearer {token}\".",
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IReturnProject"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"reason": {
											"properties": {},
											"additionalProperties": {},
											"type": "object"
										}
									},
									"required": [
										"message",
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Create a new project, with the fields \"title\", \"description\" and an optional \"open\". The title and description are registered within the project metadata (as rdfs:label and rdfs:comment), while the \"open\" parameter is used to create an initial Access Control graph (\"open\": true means that every agent has Read access to the project). Authentication is mandatory (default); the user is authenticated via a Bearer token sent along with the request as a header \"Authorization: Bearer {token}\".",
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateProject"
							}
						}
					}
				}
			}
		},
		"/lbd/public": {
			"get": {
				"operationId": "GetPublicProjects",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IReturnProject"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"reason": {
											"properties": {},
											"additionalProperties": {},
											"type": "object"
										}
									},
									"required": [
										"message",
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Retrieves the public (openly accessible) projects on the local LBDserver instance. These are, among others, projects that have been created with the property \"open\": true (and for which the ACL graph has not been modified to make it closed again). User authentication is NOT required.",
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Accept",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/lbd/{projectName}": {
			"get": {
				"operationId": "GetOneProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IReturnProject"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"reason": {
											"properties": {},
											"additionalProperties": {},
											"type": "object"
										}
									},
									"required": [
										"message",
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get a specific project by its ID (i.e. using the project URL). The user is authenticated via a Bearer token sent along with the request as a header \"Authorization: Bearer {token}\". Whether authentication is mandatory depends on the ACL policy of the requested resource. (e.g. an open project may grant READ access to an unauthenticated agent).",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectName",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Accept",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateOneProject",
				"responses": {
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"reason": {
											"properties": {},
											"additionalProperties": {},
											"type": "object"
										}
									},
									"required": [
										"message",
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get a specific project by its ID (i.e. using the project URL). The user is authenticated via a Bearer token sent along with the request as a header \"Authorization: Bearer {token}\". Whether authentication is mandatory depends on the ACL policy of the requested resource. (e.g. an open project may grant READ access to an unauthenticated agent).",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectName",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "update",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteProject",
				"responses": {
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"reason": {
											"properties": {},
											"additionalProperties": {},
											"type": "object"
										}
									},
									"required": [
										"message",
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete a project by its ID (i.e. using the project URL). Authentication is mandatory; the user is authenticated via a Bearer token sent along with the request as a header \"Authorization: Bearer {token}\".",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectName",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Accept",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/lbd/{projectName}/files/{fileId}": {
			"get": {
				"operationId": "GetOneFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IReturnResource"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"reason": {
											"properties": {},
											"additionalProperties": {},
											"type": "object"
										}
									},
									"required": [
										"message",
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Get a project File by its ID (i.e. via the File URL). The user is authenticated via a Bearer token sent along with the request as a header \"Authorization: Bearer {token}\". Whether authentication is mandatory depends on the ACL policy of the requested resource. (e.g. an open project may grant READ access to an unauthenticated agent).",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectName",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "fileId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteOneFile",
				"responses": {
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"reason": {
											"properties": {},
											"additionalProperties": {},
											"type": "object"
										}
									},
									"required": [
										"message",
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Delete a project resource (non RDF) by its ID (i.e. using the project URL). Authentication is mandatory (default); the user is authenticated via a Bearer token sent along with the request as a header \"Authorization: Bearer {token}\".",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectName",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "fileId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/lbd/{projectName}/files": {
			"post": {
				"operationId": "CreateNewFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IReturnResource"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"reason": {
											"properties": {},
											"additionalProperties": {},
											"type": "object"
										}
									},
									"required": [
										"message",
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Upload a new document (non-RDF resource) to the LBDserver. A UUID is generated for the document, its full URI is {projectURI}/files/{fileID}. A metadata graph (RDF!) gets created at {projectURI}/files/{fileID}.meta. The file should be sent as FormData with, with fileName \"resource\". The \"label\" and the \"description\" (body: as strings) are stored in the metadata graph, as well as a reference to the Access Control graph that applies to the new resource. If no ACL is specified, the metadata points towards the default project ACL (which was created at project setup). When creating a new file, a user may choose to upload a specific ACL graph along with the file itself. The ACL graph then is a FormData file identified by the fileName \"acl\". A final option is to point at an already existing ACL graph in the project, by including a (string) reference in the request body \"acl\". Note that there is currently no validation to check is an ACL graph is valid. Authentication is mandatory (default); the user is authenticated via a Bearer token sent along with the request as a header \"Authorization: Bearer {token}\".",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectName",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/lbd/{projectName}/graphs": {
			"post": {
				"operationId": "CreateNewGraph",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IReturnResource"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										},
										"reason": {
											"properties": {},
											"additionalProperties": {},
											"type": "object"
										}
									},
									"required": [
										"message",
										"reason"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Upload a new graph (RDF resource) to the LBDserver. A UUID is generated for the graph, its full URI is {projectURI}/graphs/{graphID}. A metadata graph (RDF) gets created at {projectURI}/graph/{graph}.meta. The graph should be sent as FormData with, with fileName \"resource\". The \"label\" and the \"description\" (body: as strings) are stored in the metadata graph, as well as a reference to the Access Control graph that applies to the new resource. If no ACL is specified, the metadata points towards the default project ACL (which was created at project setup). When creating a new file, a user may choose to upload a specific ACL graph along with the file itself. The ACL graph then is a FormData file identified by the fileName \"acl\". A final option is to point at an already existing ACL graph in the project, by including a (string) reference in the request body \"acl\". Note that there is currently no validation to check is an ACL graph is valid. Authentication is mandatory (default); the user is authenticated via a Bearer token sent along with the request as a header \"Authorization: Bearer {token}\".",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectName",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}